+++ web-wallet/src/lib/dusk/polyfill/index.js
@@ -2,5 +2,3 @@
 
 import "./asyncIterator";
 import "./promiseWithResolvers";
-
-import "web-streams-polyfill/polyfill";