+++ web-wallet/src/routes/(welcome)/setup/restore/MnemonicAuthenticate.svelte
@@ -1,10 +1,9 @@
 <script>
-  import { validateMnemonic, wordlists } from "bip39";
   import { mdiAlertOutline } from "@mdi/js";
 
+  import { validateMnemonic } from "$lib/wallet";
   import { Mnemonic } from "$lib/dusk/components";
   import { IconHeadingCard } from "$lib/containers/Cards";
-
   import { toast } from "$lib/dusk/components/Toast/store";
 
   /** @type {boolean} */
@@ -16,10 +15,7 @@
   /** @type {string[]} */
   export let enteredMnemonicPhrase = [];
 
-  $: isValid = validateMnemonic(
-    enteredMnemonicPhrase.join(" "),
-    wordlists.english
-  );
+  $: isValid = validateMnemonic(enteredMnemonicPhrase.join(" "));
   $: if (
     enteredMnemonicPhrase.filter((word) => word !== "").length === wordLimit &&
     !isValid