+++ web-wallet/vite.config.js
@@ -2,7 +2,6 @@ import { sveltekit } from "@sveltejs/kit/vite";
 import { coverageConfigDefaults } from "vitest/config";
 import { default as basicSsl } from "@vitejs/plugin-basic-ssl";
 import { defineConfig, loadEnv } from "vite";
-import { nodePolyfills } from "vite-plugin-node-polyfills";
 import { execSync } from "child_process";
 import { resolve } from "path";
 
@@ -14,10 +13,7 @@ export default defineConfig(({ mode }) => {
   );
   const APP_VERSION = process.env.npm_package_version ?? "unknown";
   const APP_BUILD_INFO = `${buildHash.toString() || "unknown"} ${buildDate}`;
-  const commonPlugins = [
-    sveltekit(),
-    nodePolyfills({ globals: { Buffer: true }, include: ["buffer"] }),
-  ];
+  const commonPlugins = [sveltekit()];
 
   // needed to use %sveltekit.env.PUBLIC_APP_VERSION% in app.html
   process.env.PUBLIC_APP_VERSION = APP_VERSION;