+++ web-wallet/CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Update Transactions list design [#1922]
 - Withdrawal pending message explains status and shows finalization estimate [#3954]
 - Update dependencies to address potential security issues [#3982]
+- Replace `bip39` with `@scure/bip39`, which is an audited lib, for improved security [#3984]
 
 ### Removed
 
@@ -623,6 +624,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#3955]: https://github.com/dusk-network/rusk/issues/3955
 [#3958]: https://github.com/dusk-network/rusk/issues/3958
 [#3982]: https://github.com/dusk-network/rusk/issues/3982
+[#3984]: https://github.com/dusk-network/rusk/issues/3984
 
 <!-- VERSIONS -->